Expr: Introduce Values
Rather than passing Strings to `==` etc. (remember, that's our `==`, not
the Prelude's), we introduce a new type `Value` to give it. Currently,
the only constructor is `Const`, but this is of course a step towards
having variables as well.

However, there still needs to be some work to make this work nicely.
I'm thinking perhaps a type class that can give our `==` the ability to
"automatically" turn Strings into consts, or something like that. (Large
amounts of hand-waving here.)