Expr: have expr return Maybe Prim.Boolean
The idea here is that this is what we need when we can have variable
names in the input (later on) that are not bound to a value. In that
case, we can't evaluate the expression, so we evaluate to `Nothing` as
an error indication.

However, this gets a bit annoying in the case where we deal with
multiple Exprs in a particular eval pattern match, because we have
to check each one for Nothing as output, and pass it through if that
happens.

It's not so bad for `eval e (Not x)` (as is already done in this commit)
when there's only the one, but when there are two, as with eval on
`||`, it becomes more difficult, and this is where we want to use some
clever algebraic structure thingie and lift all this into the Maybe
Monoid (well, probably Semigroup in PureScript) to make it happen
automatically.

And this is all on a branch because probably half of what I've said
here is rubbish, and I want to make sure this commit and branch will
eventually vanish forever, with nobody able to see what I've done here.